moved both search methods into APIcontroller and moved tests for these methods into APIControllerTest
