+++ src/main/APIController.java
@@ -30,4 +30,30 @@ public class APIController {
 		System.out.println("Response Body is =>  " + responseBody);
 	}
 	
+	public String searchByKeyword(String keywordInput) {
+		//call the actual search method here 
+		RestAssured.baseURI = "http://webknox.com/api/recipes";
+		String apiKey = "48ec8d46512e4c03a4d2c18b015d64af"; //Aidan API key
+		
+		RequestSpecification httpRequest = RestAssured.given();
+		Response response = httpRequest.request(Method.GET, "/complexSearch?apiKey=" + apiKey + "&query=" + keywordInput);
+		
+		String responseBody = response.getBody().asString();
+		System.out.println("Response Body is =>  " + responseBody);
+		
+		return responseBody;
+	}
+	
+	public String searchByIngredient(String ingredientInput) {
+		RestAssured.baseURI = "http://webknox.com/api/recipes";
+		String apiKey = "48ec8d46512e4c03a4d2c18b015d64af"; //Aidan API key
+		
+		RequestSpecification httpRequest = RestAssured.given();
+		Response response = httpRequest.request(Method.GET, "/findByIngredients?apiKey=" + apiKey + "&ingredients=" + ingredientInput);
+		
+		String responseBody = response.getBody().asString();
+		System.out.println("Response Body is =>  " + responseBody);
+		return responseBody;
+	}
+
 }
